# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Amiri Phone - Algeria Ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: wilaya
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
        slug:
          type: string
        icon:
          type: string
          nullable: true
        productCount:
          type: integer
      required: [id, name, slug, productCount]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        slug:
          type: string
        icon:
          type: string
      required: [name, slug]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerPhone2:
          type: string
          nullable: true
        wilayaCode:
          type: string
        wilayaName:
          type: string
        communeCode:
          type: string
          nullable: true
        communeName:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, returned, cancelled]
        notes:
          type: string
          nullable: true
        deliveryCompany:
          type: string
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, orderNumber, customerName, customerPhone, wilayaCode, wilayaName, items, subtotal, shippingCost, total, status, createdAt, updatedAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
          nullable: true
        quantity:
          type: integer
        price:
          type: number
        total:
          type: number
      required: [id, productId, productName, quantity, price, total]
    CreateOrderRequest:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerPhone2:
          type: string
        wilayaCode:
          type: string
        wilayaName:
          type: string
        communeCode:
          type: string
        communeName:
          type: string
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemRequest"
        notes:
          type: string
      required: [customerName, customerPhone, wilayaCode, wilayaName, items]
    CreateOrderItemRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, returned, cancelled]
        notes:
          type: string
        deliveryCompany:
          type: string
        trackingNumber:
          type: string
      required: [status]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [orders, total, limit, offset]
...